+++ samha.backend/src/main/java/com/samha/persistence/IUsuarioRepository.java
@@ -4,6 +4,6 @@ package com.samha.persistence;
 import com.samha.domain.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IUsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface IUsuarioRepository extends JpaRepository<Usuario, Long> {
     Usuario findByLogin(String login);
 }
